+++ src/server/client_write.c
@@ -2,7 +2,7 @@
 #include <sys/socket.h>
 #include "irc.h"
 
-void	client_write(t_env *e, int cs)
+void	client_write(t_env *e, size_t cs)
 {
     (void)e;
     (void)cs;
